Reorganize maintenance notification code
