+++ frontend/src/pages/AuthLogin/AuthLogin.tsx
@@ -3,9 +3,27 @@ import { useAuthContext } from 'context';
 import { Button } from '@trussworks/react-uswds';
 import { Alert, AlertTitle, Box, Grid, Typography } from '@mui/material';
 import { CrossfeedWarning } from 'components/WarningBanner';
-import { initialNotificationValues, MaintenanceNotification } from 'types';
+import { MaintenanceNotification } from 'types';
 import { v4 as uuidv4 } from 'uuid';
 import pkceChallenge from 'pkce-challenge';
+
+const MaintenanceAlert: React.FC<any> = ({ notification }) => {
+  if (!notification) return null;
+  const isLoginUnavailable =
+    notification?.maintenance_type === 'major' &&
+    notification?.status === 'active';
+  const titleText = isLoginUnavailable
+    ? 'CyHy Dashboard Major Maintenance: Login Not Available'
+    : 'CyHy Dashboard Maintenance Notification';
+  return (
+    <Grid size={{ xs: 12 }}>
+      <Alert severity="warning">
+        <AlertTitle>{titleText}</AlertTitle>
+        {notification?.message}
+      </Alert>
+    </Grid>
+  );
+};
 
 const LoginButton = () => {
   // TODO: Capture default values here once determined
@@ -18,11 +36,6 @@ const LoginButton = () => {
     const { code_challenge, code_verifier } = await pkceChallenge();
     const state = uuidv4();
 
-    console.log('Starting OAuth fetch with:', {
-      code_challenge,
-      code_verifier,
-      state
-    });
     try {
       const res = await fetch(
         `${import.meta.env.VITE_API_URL}/auth/get-oauth-meta`,
@@ -35,7 +48,6 @@ const LoginButton = () => {
 
       const json = await res.json();
       localStorage.setItem('oauthMeta', json.signedToken);
-      console.log('Stored oauthMeta:', json.signedToken);
 
       const authorizeUrl = `https://${domain}/oauth2/authorize?identity_provider=Okta&redirect_uri=${encodeURIComponent(
         callbackUrl
@@ -64,7 +76,7 @@ const LoginButton = () => {
 export const AuthLogin: React.FC<{ showSignUp?: boolean }> = () => {
   const { apiGet } = useAuthContext();
   const [notification, setNotification] =
-    React.useState<MaintenanceNotification>(initialNotificationValues);
+    React.useState<MaintenanceNotification | null>(null);
   const fetchNotifications = React.useCallback(async () => {
     try {
       const rows = await apiGet('/notifications');
@@ -86,28 +98,9 @@ export const AuthLogin: React.FC<{ showSignUp?: boolean }> = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const MaintenanceAlert: React.FC<any> = ({ notification }) => {
-    // Determine the conditional title
-    const isLoginUnavailable =
-      notification?.maintenance_type === 'major' &&
-      notification?.status === 'active';
-    const titleText = isLoginUnavailable
-      ? 'CyHy Dashboard Major Maintenance: Login Not Available'
-      : 'CyHy Dashboard Maintenance Notification';
-
-    return <AlertTitle>{titleText}</AlertTitle>;
-  };
-
-  const platformNotification = (
-    <Grid size={{ xs: 12 }}>
-      <Alert severity="warning">
-        <MaintenanceAlert notification={notification} />
-        {notification?.message}
-      </Alert>
-    </Grid>
-  );
   return (
     <Box display="flex" flexDirection="column" height={'calc(100vh - 108px)'}>
+      {notification && <MaintenanceAlert notification={notification} />}
       <Box flex={0.5} display="flex" />
       <Box
         flex={0.5}
