+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -5,6 +5,7 @@ import { Box, BoxProps, Grid, CircularProgress, Link } from '@mui/material';
 import { ContextType, useAuthContext } from 'context';
 import { useVulnScanData } from 'hooks/useVulnScanData';
 import { useCheckUserState } from 'hooks/useCheckUserState';
+import { useMaintenanceNotifications } from 'hooks/useMaintenanceNotifications';
 // Constants & types
 import { ORGANIZATION_FILTER_KEY } from 'components/FilterDrawer/VSDashRegionAndOrgFilters';
 import { OrganizationShallow } from 'components/FilterDrawer/RegionAndOrganizationFilters';
@@ -71,55 +72,12 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
     user,
     isLoggingOut
   );
-  const [isLoginBlockedDialogOpen, setIsLoginBlockedDialogOpen] =
-    useState(false);
-  const [maintenanceNotification, setMaintenanceNotification] =
-    useState<any>(null);
-
-  useEffect(() => {
-    const fetchAndCheckMaintenance = async () => {
-      if (
-        user &&
-        !user.invite_pending &&
-        user.state &&
-        user.date_accepted_terms &&
-        !isLoginBlockedDialogOpen
-      ) {
-        const notifications = await apiGet('/notifications');
-        const active = notifications.find(
-          (n: any) =>
-            n.status === 'active' &&
-            n.maintenance_type === 'major' &&
-            new Date(n.start_datetime) <= new Date() &&
-            new Date(n.end_datetime) >= new Date()
-        );
-        const nonBlockingUserTypes = ['globalAdmin', 'regionalAdmin'];
-        if (active && !nonBlockingUserTypes.includes(user.user_type)) {
-          setMaintenanceNotification(active);
-          setIsLoginBlockedDialogOpen(true);
-        }
-      }
-    };
-
-    fetchAndCheckMaintenance();
-  }, [apiGet, isLoginBlockedDialogOpen, user, userMustSign]);
-
-  useEffect(() => {
-    const handleMaintenanceBlocked = (e: any) => {
-      if (e.detail?.message) {
-        setMaintenanceNotification({ message: e.detail.message });
-        setIsLoginBlockedDialogOpen(true);
-      }
-    };
-
-    window.addEventListener('maintenance-blocked', handleMaintenanceBlocked);
-    return () => {
-      window.removeEventListener(
-        'maintenance-blocked',
-        handleMaintenanceBlocked
-      );
-    };
-  }, []);
+  const {
+    isLoginBlockedDialogOpen,
+    setIsLoginBlockedDialogOpen,
+    maintenanceNotification,
+    setMaintenanceNotification
+  } = useMaintenanceNotifications(user, apiGet, userMustSign);
 
   useEffect(() => {
     filters.forEach((filter) => {
